requirements.txt
aiofiles==24.1.0
astor==0.8.1
blinker==1.6.2
certifi==2023.11.17
charset-normalizer==3.3.2
chess==1.10.0
click==8.1.7
contourpy==1.2.0
coverage==7.4.0
cycler==0.12.1
dnspython==2.5.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
fonttools==4.48.1
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.4
oauthlib==3.2.2
packaging==23.1
# pandas==2.2.0
pillow==10.2.0
priority==2.0.0
pyarrow==15.0.0
# pycairo @ file:///private/tmp/py3cairo-20240213-5896-uig1tg/pycairo-1.26.0
pydocstyle==6.3.0
PyJWT==2.8.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
Quart==0.19.6
quart-cors==0.7.0
requests==2.31.0
six==1.16.0
snowballstemmer==2.2.0
tzdata==2024.1
urllib3==2.1.0
uvicorn==0.30.6
Werkzeug==3.0.4
wsproto==1.2.0
